Don't reinitialize seed after it has been deleted
On the initial topology discovery, it is possible that the mongoDB list
of servers differs from the list of seeds (due to, for instance,
specifying seed locations as ip addresses that then get resolved to
hostnames). If that is the case, the seed is marked as deleted but we keep
trying to initialize it. This commit makes it so that we stop trying to
initialize the removed seed, and keep the determined list of servers.
It should prevent an infinite startup loop.